# UTFPR/Web/web2-hotel/backend/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Usuários",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/usuarios": {
      "get": {
        "summary": "Lista todos os usuários",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de usuários"
          }
        }
      },
      "post": {
        "summary": "Cria um novo usuário",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": { "type": "string" },
                  "nascimento": { "type": "string", "format": "date" },
                  "email": { "type": "string" },
                  "senha": { "type": "string" },
                  "perfil": { "type": "string" }
                },
                "required": ["nome", "email", "senha", "perfil"]
              }
            }
          },
          "responses": {
            "201": {
              "description": "Usuário criado com sucesso"
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}